Use String#each_char method.
Defining the String#each can be dangerous beacuse this implementations
is not as performant as the C implementation of #each_char.

For example: by only including ruby_figlet in a Rails application, the
transfer of the response is slowed down by some orders of magintude (from
few millis to many seconds).